+++ src/Lwt/Interfaces/ITextRepository.cs
@@ -26,6 +26,6 @@ namespace Lwt.Interfaces
         /// <param name="userId">the user id.</param>
         /// <param name="textFilter">the filters.</param>
         /// <returns>A <see cref="Task{TResult}"/> representing the result of the asynchronous operation.</returns>
-        Task<int> CountByUserAsync(Guid userId, TextFilter textFilter);
+        Task<long> CountByUserAsync(Guid userId, TextFilter textFilter);
     }
 }
