+++ src/Lwt/Interfaces/Services/ITextService.cs
@@ -50,6 +50,6 @@ namespace Lwt.Interfaces.Services
         /// <param name="userId">the user id.</param>
         /// <param name="textFilter">the filters.</param>
         /// <returns>A <see cref="Task{TResult}"/> representing the result of the asynchronous operation.</returns>
-        Task<int> CountAsync(Guid userId, TextFilter textFilter);
+        Task<long> CountAsync(Guid userId, TextFilter textFilter);
     }
 }
